Rename to core_database
