+++ src/python/ensembl/io/genomio/database/factory.py
@@ -29,7 +29,7 @@ from sqlalchemy.engine import URL
 from ensembl.utils.argparse import ArgumentParser
 from ensembl.utils.logging import init_logging_with_args
 from .core_server import CoreServer
-from .core_db import CoreDatabase
+from .core_database import CoreDatabase
 
 
 _DB_PATTERN = re.compile(r".+_core_(\d+)_\d+_\d+")
